fix: type safety improvements for Android pipeline
Fix unsigned long handling and type conversions in Kotlin port
of the Rust detection pipeline.

Changes:
- RouteDataParser: Proper ULong conversion for bitmask parsing
- MapMatcher: ULong bitwise operations for grid cell matching
- SemanticTypes: Safe Short conversion via Int
- AndroidManifest: Add Google Maps API key meta-data

These fixes ensure the Kotlin pipeline matches the Rust implementation's
behavior for unsigned integer operations.

+++ android/app/src/main/java/com/busarrival/app/data/pipeline/binary/RouteDataParser.kt
@@ -64,7 +64,7 @@ object RouteDataParser {
         // Verify CRC32
         val crcStored = buffer.int
         val crcCalc = calculateCrc(bytes, 9) // Skip magic (4) + version (1) + crc (4)
-        require(crcCalc == crcStored.toLong()) { "CRC mismatch: expected $crcStored, got $crcCalc" }
+        require(crcCalc.toLong() == crcStored.toLong()) { "CRC mismatch: expected $crcStored, got $crcCalc" }
 
         // Read header
         val originLat = buffer.int
@@ -127,7 +127,7 @@ object RouteDataParser {
         val totalCells = rows * cols
 
         repeat(totalCells) {
-            val bitmask = buffer.long
+            val bitmask = buffer.long.toULong()
             val offsetCount = buffer.get().toInt() and 0xFF
 
             val offsets = mutableListOf<Int>()
